% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid_peptides.R
\docType{data}
\name{Covid_peptides}
\alias{Covid_peptides}
\title{Large-scale proteomic Analysis of COVID-19 Severity}
\format{
A dataframe with 1447720 rows and 6 variables:
\describe{
\item{Peptide.Modified.Sequence}{Peptide sequence with modifications}
\item{Protein}{Name of protein}
\item{value}{Intensity values }
\item{Condition}{Experiment Condition}
\item{BioReplicate}{Name of Biological Replicate}
}
}
\source{
\url{https://github.com/jessegmeyerlab/PeCorA/blob/master/inputs/covid_peptides.zip}
}
\usage{
Covid_peptides
}
\description{
Shotgun proteomics of COVID-19 ARDS patients from Overmyera et al 2020.
}
\keyword{datasets}
