% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_processed_data_for_PeCorA.R
\name{import_processed_data_for_PeCorA}
\alias{import_processed_data_for_PeCorA}
\title{import_processed_data_for_PeCorA}
\usage{
import_processed_data_for_PeCorA(x)
}
\arguments{
\item{x}{peptide.txt MaxQuant output file}
}
\value{
dataframe ready for PeCorA analysis
}
\description{
reads peptide.txt MaxQuant output example file inside the package and process them for PeCorA
}
\details{
output files containing columns: Peptide.Modified.Sequence", Protein, variable,value,Condition,BioReplicate
}
\examples{
\dontrun{
if(interactive()){
t<-import_processed_data_for_PeCorA("peptides.txt")
 }
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
