% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_processed_data_for_PeCorA_LFQ.R
\name{import_processed_data_for_PeCorA_LFQ}
\alias{import_processed_data_for_PeCorA_LFQ}
\title{import_processed_data_for_PeCorA_LFQ}
\usage{
import_processed_data_for_PeCorA_LFQ(
  LFQfile,
  condition1,
  condition2,
  condition3
)
}
\arguments{
\item{LFQfile}{peptide.txt MaxQuant output file}

\item{condition1}{experimental condition 1}

\item{condition2}{experimental condition 2}

\item{condition3}{experimental condition 3}
}
\value{
dataframe ready for PeCorA analysis
}
\description{
reads peptide.txt MaxQuant output example file inside the package and process them for PeCorA.
}
\details{
output files containing columns: Peptide.Modified.Sequence", Protein, variable,value,Condition,BioReplicate
}
\examples{
\dontrun{
if(interactive()){
t <- import_processed_data_for_PeCorA_LFQ( LFQfile = "peptides.txt",condition1="CONTROL",condition2="_COVID",condition3="NON.COVID")
 }
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
