% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA_alldf.R
\name{PeCorA_alldf}
\alias{PeCorA_alldf}
\title{PeCorA_alldf}
\usage{
PeCorA_alldf(t)
}
\arguments{
\item{t}{scale data}
}
\value{
alldf dataframe output of PeCorA
}
\description{
output alldf of PeCorA
}
\details{
describe the columns
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
