% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA_noZ.R
\docType{data}
\name{PeCorA_noZ}
\alias{PeCorA_noZ}
\title{Proteomic analysis of alcohol and LPS-treated primary mouse microglia}
\format{
A dataframe with 415275 rows and 8 variables:
\describe{
\item{Peptide}{Peptide sequence}
\item{Protein}{Name of protein}
\item{Peptide.Modified.Sequence}{Peptide sequence with modifications}
\item{Begin.Pos}{Position of peptide begining}
\item{End.Pos}{Position of peptide end}
\item{Condition}{Experiment Condition}
\item{BioReplicate}{Number of Biological Replicate}
\item{Normalized.Area}{Value of normalized areas}
}
}
\source{
\url{https://github.com/jessegmeyerlab/PeCorA/blob/master/inputs/PeCorA_noZ.zip}
}
\usage{
PeCorA_noZ
}
\description{
Mouse microglia proteome processed with Skyline from Guergues et al 2020.
}
\keyword{datasets}
