% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA_plotting.R
\name{PeCorA_plotting}
\alias{PeCorA_plotting}
\title{PeCorA_plotting}
\usage{
PeCorA_plotting(
  disagree_peptides,
  disagree_peptides_selection,
  scaled_peptides
)
}
\arguments{
\item{disagree_peptides}{dataframe object with disagree peptides}

\item{disagree_peptides_selection}{dataframe object selection of disagree peptides}

\item{scaled_peptides}{scaled peptides}
}
\value{
PeCorA plot
}
\description{
generates boxplot using PeCorA output
}
\details{
This function generates boxplots for a given protein representing
the peptides that are statistically different from the quantities of all the
other peptides in green and all other peptides in grey.
}
\examples{
if (interactive()) {
  PeCorA_plotting_plot <- PeCorA_plotting(
    disagree_peptides,
    disagree_peptides[12, ], scaled_peptides
  )
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
