% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA_preprocessing.R
\name{PeCorA_preprocessing}
\alias{PeCorA_preprocessing}
\title{PeCorA_preprocessing}
\usage{
PeCorA_preprocessing(
  pecora_format,
  area_column_name,
  threshold_to_filter,
  control_name
)
}
\arguments{
\item{pecora_format}{dataframe in PeCorA format}

\item{area_column_name}{numeric column number with peak areas}

\item{threshold_to_filter}{numeric threshold for peak areas to filter}

\item{control_name}{character control reference}
}
\value{
dataframe output of pre-processed data ready for PeCorA analysis
}
\description{
steps to pre-process data for PeCorA
}
\details{
Peak areas are log transformed and scaled to center.
After global scaling, each peptide scaling was performed to center
each peptide relative to the mean of the control group’s peak area.
}
\examples{
if(interactive()){
scaled_peptides <- PeCorA_preprocessing(pecora_format,
                                         area_column_name=4,
                                         threshold_to_filter=100,
                                         control_name="control")
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
