% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptides_data_filtered}
\alias{peptides_data_filtered}
\title{COVID-19 plasma proteomics}
\format{
A dataframe with 10645 rows and 169 variables:
\describe{
\item{Sequence}{Peptide sequence}
\item{Leading.razor.protein}{Name of protein}
\item{LFQ.intensity columns (136)}{LFQ normalized mass spectrometry
intensity, A.U.}
}
}
\source{
Overmyer et al 2020.
Large-Scale Multi-omic Analysis of COVID-19 Severity
https://doi.org/10.1016/j.cels.2020.10.003
\href{https://www.sciencedirect.com/science/article/pii/S2405471220303719}
{doi: 10.1016/j.cels.2020.10.003}.
}
\usage{
peptides_data_filtered
}
\value{
A data.frame
}
\description{
Large-Scale Multi-omic Analysis of COVID-19 Severity processed with MaxQuant
from Overmyer et al 2020.
The dataset contains the peptide output file from
\href{http://www.maxquant.org}{MaxQuant}.
}
\keyword{datasets}
