% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_by.R
\name{scale_by}
\alias{scale_by}
\title{scale_by}
\usage{
scale_by(object = NULL, data = NULL, scale = 1)
}
\arguments{
\item{object}{NULL}

\item{data}{NULL}

\item{scale}{NULL}
}
\value{
center and scales a numeric variable within each level of a factor
(or the interaction of several factors
}
\description{
scale_by
}
\examples{
if (interactive()) {
  dat <- data.frame(
  f1 = rep(c("a", "b", "c"), c(5, 10, 20)),
  x1 = rnorm(35, rep(c(1, 2, 3), c(5, 10, 20)),
  rep(c(.5, 1.5, 3), c(5, 10, 20))))
  dat$x1_scaled_by_f1 <- scale_by(x1 ~ f1, dat)
}
}
