\name{signFlip-package}
\alias{signFlip-package}
\alias{signFlip}
\docType{package}
\title{
\packageTitle{signFlip}
}
\description{
\packageDescription{signFlip}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{signFlip}
\packageIndices{signFlip}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{signFlip}

Maintainer: \packageMaintainer{signFlip}
}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }

