% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{splt_last}
\alias{splt_last}
\title{splt_last}
\usage{
splt_last(x)
}
\arguments{
\item{x}{character of names}
}
\value{
list with number followed by underscore
}
\description{
splt_last
}
\examples{
if (interactive()) {
  lapply(FUN = splt_last, paste0(list("apple", "pear"), "_1"))
}
}
