% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2p.R
\name{t2p}
\alias{t2p}
\title{From test statistics to p-values}
\usage{
t2p(Test_matrix, alternative, exact = TRUE, permReturn, gdl = NULL)
}
\arguments{
\item{Test_matrix}{matrix where columns represent the B permutations and rows the m tests statistic. 
The observed test statistic is in the first column
and the permutation distribution in the remaining columns}

\item{alternative}{character string referring to the alternative hypothesis (\code{"greater"}, \code{"lower"}, or \code{"two.sided"}).}

\item{exact}{logical value, \code{TRUE} to compute p-values by permutation distribution. Default @TRUE.}

\item{permReturn}{logical value, \code{TRUE} to return the t-tests and p-values permutation distribution.}

\item{gdl}{numerical value. Degree of freedom.}
}
\value{
matrix of p-values where columns represent the B permutations and rows the m tests statistic.
}
\description{
Use permutation distribution of a test statistic to get p-values.
}
