% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA.R
\name{PeCorA}
\alias{PeCorA}
\title{PeCorA}
\usage{
PeCorA(t)
}
\arguments{
\item{t}{scaled data}
}
\value{
dataframe output of PeCorA analysis containing values for proteins/peptides adj_pval for peptide interaction across condition and peptide groups.
}
\description{
core function of PeCorA workflow
}
\details{
PeCorA function relies on linear model to assess whether the slope of a peptide’s change across treatment groups differs from the slope of all other peptides assigned to the same protein.
}
\examples{
\dontrun{
if(interactive()){
 disagree_peptides <- PeCorA (scaled_peptides)
 }
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
