% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA_preprocessing.R
\name{PeCorA_preprocessing}
\alias{PeCorA_preprocessing}
\title{PeCorA_preprocessing}
\usage{
PeCorA_preprocessing(t, area_column_name, threshold_to_filter, control_name)
}
\arguments{
\item{t}{data in PeCorA format}

\item{area_column_name}{column number with peak areas}

\item{threshold_to_filter}{threshould value of peak areas to filter}

\item{control_name}{control reference}
}
\value{
dataframe output of pre-process data ready for PeCorA analalysis
}
\description{
steps to pre-process data for PeCorA
}
\details{
Peak areas are log transformed and scaled to center.After global scaling, each peptide scaling was performed to center each peptide relative to the mean of the control group’s peak area.
}
\examples{
\dontrun{
if(interactive()){
 scaled_peptides <- PeCorA_preprocessing(t)
 }
}
}
\author{
Maria Dermit \href{mailto:maria.dermit@qmul.ac.uk}{maria.dermit@qmul.ac.uk}
}
